+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/hints/ParadoxLocalisationReferenceHintsProvider.kt
@@ -20,7 +20,7 @@ import javax.swing.*
 class ParadoxLocalisationReferenceHintsProvider : ParadoxLocalisationHintsProvider<Settings>() {
     data class Settings(
         var textLengthLimit: Int = 30,
-        var iconHeightLimit: Int = 32
+        var iconHeightLimit: Int = 36
     )
 
     private val settingsKey = SettingsKey<Settings>("ParadoxLocalisationReferenceHintsSettingsKey")