+++ src/main/kotlin/icu/windea/pls/script/codeInsight/hints/ParadoxComplexEnumValueLocalizedNameHintsProvider.kt
@@ -25,7 +25,7 @@ import javax.swing.*
 class ParadoxComplexEnumValueLocalizedNameHintsProvider : ParadoxScriptHintsProvider<Settings>() {
     data class Settings(
         var textLengthLimit: Int = 30,
-        var iconHeightLimit: Int = 32
+        var iconHeightLimit: Int = 36
     )
 
     private val settingsKey = SettingsKey<Settings>("ParadoxComplexEnumValueLocalizedNameHintsSettingsKey")