+++ src/main/kotlin/icu/windea/pls/script/codeInsight/hints/ParadoxDefinitionReferenceLocalizedNameHintsProvider.kt
@@ -23,7 +23,7 @@ import javax.swing.*
 class ParadoxDefinitionReferenceLocalizedNameHintsProvider : ParadoxScriptHintsProvider<Settings>() {
     data class Settings(
         var textLengthLimit: Int = 30,
-        var iconHeightLimit: Int = 32
+        var iconHeightLimit: Int = 36
     )
 
     private val settingsKey = SettingsKey<Settings>("ParadoxDefinitionReferenceLocalizedNameHintsSettingsKey")