+++ src/main/kotlin/icu/windea/pls/script/codeInsight/hints/ParadoxLocalisationReferenceInfoHintsProvider.kt
@@ -24,7 +24,7 @@ import javax.swing.*
 class ParadoxLocalisationReferenceInfoHintsProvider : ParadoxScriptHintsProvider<Settings>() {
     data class Settings(
         var textLengthLimit: Int = 30,
-        var iconHeightLimit: Int = 32
+        var iconHeightLimit: Int = 36
     )
 
     private val settingsKey = SettingsKey<Settings>("ParadoxLocalisationReferenceInfoHintsSettingsKey")