+++ src/main/kotlin/icu/windea/pls/script/codeInsight/hints/ParadoxModifierLocalizedNameHintsProvider.kt
@@ -27,7 +27,7 @@ import javax.swing.*
 class ParadoxModifierLocalizedNameHintsProvider : ParadoxScriptHintsProvider<Settings>() {
     data class Settings(
         var textLengthLimit: Int = 30,
-        var iconHeightLimit: Int = 32
+        var iconHeightLimit: Int = 36
     )
 
     private val settingsKey = SettingsKey<Settings>("ParadoxModifierLocalizedNameHintsSettingsKey")